Frontend design for staff page
Front end designed to look like wire frames
